+++ src/main/java/neofontrender/core/font/FontManager.java
@@ -212,6 +212,7 @@ public class FontManager implements AutoCloseable {
                 NeofontrenderConfig.fontAntialiasMode(),
                 NeofontrenderConfig.fontFractionalMetrics(),
                 NeofontrenderConfig.fontStyle(),
+                NeofontrenderConfig.fontVariableWeight(),
                 allowDefaultFallback
         );
     }